changes
changes

+++ ICT4Events/EventManagementSystem/Form1.cs
@@ -36,10 +36,15 @@ namespace EventManagementSystem
         {
             try
             {
-                superManager.AddEvent(Convert.ToInt32(tbEventEventID.Text), tbEventLocatie.Text,
-                    tbEventBeginDatum.Text, tbEventEindDatum.Text, tbEventBeschrijving.Text,
-                    Convert.ToDecimal(tbEventPrijs.Text));
-                MessageBox.Show("Event " + tbEventEventID.Text + " toegevoegd");
+                DialogResult result1 = MessageBox.Show("Weet u zeker dat u dit event wilt toevoegen?",
+		"Critical Warning",MessageBoxButtons.OKCancel,MessageBoxIcon.Warning);
+                if (result1 == DialogResult.Yes)
+                {
+                    superManager.AddEvent(Convert.ToInt32(tbEventEventID.Text), tbEventLocatie.Text,
+                        tbEventBeginDatum.Text, tbEventEindDatum.Text, tbEventBeschrijving.Text,
+                        Convert.ToDecimal(tbEventPrijs.Text));
+                    MessageBox.Show("Event " + tbEventEventID.Text + " toegevoegd");
+                }
             }
             catch (FormatException)
             {
@@ -59,8 +64,13 @@ namespace EventManagementSystem
         {
             try
             {
-                superManager.DeleteEvent(tbEventEventID.Text);
-                MessageBox.Show("Event " + tbEventEventID.Text + " verwijderd");
+                DialogResult result1 = MessageBox.Show("Weet u zeker dat u dit event wilt verwijderen?",
+		"Critical Warning",MessageBoxButtons.OKCancel,MessageBoxIcon.Warning);
+                if (result1 == DialogResult.Yes)
+                {
+                    superManager.DeleteEvent(tbEventEventID.Text);
+                    MessageBox.Show("Event " + tbEventEventID.Text + " verwijderd");
+                }
             }
             catch
             {
@@ -77,14 +87,23 @@ namespace EventManagementSystem
         {
             try
             {
-                superManager.EditEvent(Convert.ToInt32(tbEventEventID.Text), tbEventLocatie.Text,
-                    tbEventBeginDatum.Text, tbEventEindDatum.Text, tbEventBeschrijving.Text,
-                    Convert.ToDecimal(tbEventPrijs.Text));
-                MessageBox.Show("Event " + tbEventEventID.Text + " aangepast");
+                DialogResult result1 = MessageBox.Show("Weet u zeker dat u dit event wilt bewerken?",
+		"Critical Warning",MessageBoxButtons.OKCancel,MessageBoxIcon.Warning);
+                if (result1 == DialogResult.Yes)
+                {
+                    superManager.EditEvent(Convert.ToInt32(tbEventEventID.Text), tbEventLocatie.Text,
+                        tbEventBeginDatum.Text, tbEventEindDatum.Text, tbEventBeschrijving.Text,
+                        Convert.ToDecimal(tbEventPrijs.Text));
+                    MessageBox.Show("Event " + tbEventEventID.Text + " aangepast");
+                }
             }
-            catch
+            catch (FormatException)
             {
-
+                MessageBox.Show("Geen geldige invoer");
+            }
+            catch (MyException mex)
+            {
+                MessageBox.Show(mex.Message);
             }
             finally
             {
@@ -125,26 +144,36 @@ namespace EventManagementSystem
                 //check if it is a Guest
                 if (cbMedewerkersRole.SelectedItem.ToString() == "Guest")
                 {
-                    superManager.EditGuest(Convert.ToInt32(tbMedewerkerAccountID.Text),
-                        Convert.ToInt32(tbMedewerkersEventID.Text), tbMedewerkerUsername.Text,
-                        tbMedewerkerPassword.Text, tbMedewerkerName.Text, tbMedewerkerAddress.Text,
-                        tbMedewerkerCity.Text, tbMedewerkerPostalcode.Text, tbMedewerkerDateOfBirth.Text,
-                        tbMedewerkerEmail.Text, Convert.ToInt32(tbMedewerkerPhonenumber.Text));
-                    MessageBox.Show("Persoon " + tbMedewerkerName.Text + " aangepast");
-                    isEmployee = false;
-                    btnMedewerkersShowAllGuest_Click(sender, e);
+                    DialogResult result1 = MessageBox.Show("Weet u zeker dat u deze gast wilt bewerken?",
+		"Critical Warning",MessageBoxButtons.OKCancel,MessageBoxIcon.Warning);
+                    if (result1 == DialogResult.Yes)
+                    {
+                        superManager.EditGuest(Convert.ToInt32(tbMedewerkerAccountID.Text),
+                            Convert.ToInt32(tbMedewerkersEventID.Text), tbMedewerkerUsername.Text,
+                            tbMedewerkerPassword.Text, tbMedewerkerName.Text, tbMedewerkerAddress.Text,
+                            tbMedewerkerCity.Text, tbMedewerkerPostalcode.Text, tbMedewerkerDateOfBirth.Text,
+                            tbMedewerkerEmail.Text, Convert.ToInt32(tbMedewerkerPhonenumber.Text));
+                        MessageBox.Show("Persoon " + tbMedewerkerName.Text + " aangepast");
+                        isEmployee = false;
+                        btnMedewerkersShowAllGuest_Click(sender, e);
+                    }
                 }
                 else // it's an Employee
                 {
-                    superManager.EditEmployee(Convert.ToInt32(tbMedewerkerAccountID.Text),
+                    DialogResult result1 = MessageBox.Show("Weet u zeker dat u deze medewerker wilt?",
+		"Critical Warning",MessageBoxButtons.OKCancel,MessageBoxIcon.Warning);
+                    if (result1 == DialogResult.Yes)
+                    {
+                        superManager.EditEmployee(Convert.ToInt32(tbMedewerkerAccountID.Text),
                     Convert.ToInt32(tbMedewerkersEventID.Text), tbMedewerkerUsername.Text,
                     tbMedewerkerPassword.Text, tbMedewerkerName.Text, tbMedewerkerAddress.Text,
                     tbMedewerkerCity.Text, tbMedewerkerPostalcode.Text, tbMedewerkerDateOfBirth.Text,
                     tbMedewerkerEmail.Text, Convert.ToInt32(tbMedewerkerPhonenumber.Text),
-                    Convert.ToInt32(tbMedewerkerEmployeeID.Text), cbMedewerkersRole.SelectedIndex);
-                    MessageBox.Show("Medewerker " + tbMedewerkerName.Text + " aangepast");
-                    isEmployee = true;
-                    btnMedewerkersShowAllEmployee_Click(sender, e);
+                    Convert.ToInt32(tbMedewerkerEmployeeID.Text), cbMedewerkersRole.SelectedIndex + 1);
+                        MessageBox.Show("Medewerker " + tbMedewerkerName.Text + " aangepast");
+                        isEmployee = true;
+                        btnMedewerkersShowAllEmployee_Click(sender, e);
+                    }
                 }
             }
             catch(MyException mex)
@@ -161,13 +190,18 @@ namespace EventManagementSystem
         {
             try
             {
-                superManager.AddEmployee(Convert.ToInt32(tbMedewerkerAccountID.Text),
-                    Convert.ToInt32(tbMedewerkersEventID.Text), tbMedewerkerUsername.Text,
-                    tbMedewerkerPassword.Text, tbMedewerkerName.Text, tbMedewerkerAddress.Text,
-                    tbMedewerkerCity.Text, tbMedewerkerPostalcode.Text, tbMedewerkerDateOfBirth.Text,
-                    tbMedewerkerEmail.Text, Convert.ToInt32(tbMedewerkerPhonenumber.Text),
-                    Convert.ToInt32(tbMedewerkerEmployeeID.Text), cbMedewerkersRole.SelectedIndex + 1);
-                MessageBox.Show("Medewerker " + tbEventEventID.Text + " toegevoegd");
+                DialogResult result1 = MessageBox.Show("Weet u zeker dat u deze medewerker wilt toevoegen?",
+		"Critical Warning",MessageBoxButtons.OKCancel,MessageBoxIcon.Warning);
+                if (result1 == DialogResult.Yes)
+                {
+                    superManager.AddEmployee(Convert.ToInt32(tbMedewerkerAccountID.Text),
+                        Convert.ToInt32(tbMedewerkersEventID.Text), tbMedewerkerUsername.Text,
+                        tbMedewerkerPassword.Text, tbMedewerkerName.Text, tbMedewerkerAddress.Text,
+                        tbMedewerkerCity.Text, tbMedewerkerPostalcode.Text, tbMedewerkerDateOfBirth.Text,
+                        tbMedewerkerEmail.Text, Convert.ToInt32(tbMedewerkerPhonenumber.Text),
+                        Convert.ToInt32(tbMedewerkerEmployeeID.Text), cbMedewerkersRole.SelectedIndex + 1);
+                    MessageBox.Show("Medewerker " + tbEventEventID.Text + " toegevoegd");
+                }
             }
             catch (MyException mex)
             {
@@ -192,17 +226,26 @@ namespace EventManagementSystem
                 //check if it is a Guest
                 if (cbMedewerkersRole.SelectedItem.ToString() == "Guest")
                 {
-
-                    superManager.DeleteAccountGuest(Convert.ToInt32(tbMedewerkerAccountID.Text));
-                    btnMedewerkersShowAllGuest_Click(sender, e);
-                    isEmployee = false;
+                    DialogResult result1 = MessageBox.Show("Weet u zeker dat u deze gast wilt verwijderen?",
+		"Critical Warning",MessageBoxButtons.OKCancel,MessageBoxIcon.Warning);
+                    if (result1 == DialogResult.Yes)
+                    {
+                        superManager.DeleteAccountGuest(Convert.ToInt32(tbMedewerkerAccountID.Text));
+                        btnMedewerkersShowAllGuest_Click(sender, e);
+                        isEmployee = false;
+                    }
 
                 }
                 else // it's an Employee
                 {
-                    superManager.DeleteAccountEmployee(Convert.ToInt32(tbMedewerkerAccountID.Text));
-                    btnMedewerkersShowAllEmployee_Click(sender, e);
-                    isEmployee = true;
+                    DialogResult result1 = MessageBox.Show("Weet u zeker dat u deze medewerker wilt verwijderen?",
+		"Critical Warning",MessageBoxButtons.OKCancel,MessageBoxIcon.Warning);
+                    if (result1 == DialogResult.Yes)
+                    {
+                        superManager.DeleteAccountEmployee(Convert.ToInt32(tbMedewerkerAccountID.Text));
+                        btnMedewerkersShowAllEmployee_Click(sender, e);
+                        isEmployee = true;
+                    }
                 }
             }
             catch (MyException mex)
@@ -218,11 +261,34 @@ namespace EventManagementSystem
         //nog doen
         private void btnPlaatsEdit_Click(object sender, EventArgs e)
         {
+            try
+            {
+                DialogResult result1 = MessageBox.Show("Weet u zeker dat u deze plaats wilt aanpassen?",
+        "Critical Warning", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
+                if (result1 == DialogResult.Yes)
+                {
+                    superManager.EditLocation(Convert.ToInt32(tbPlaatsLocationID.Text), tbPlaatsName.Text,
+                        tbPlaatsDescription.Text, Convert.ToInt32(tbPlaatsPrice.Text));
+                    MessageBox.Show("Plaats " + tbPlaatsName.Text + " aangepast");
+                }
+            }
+            catch (FormatException)
+            {
+                MessageBox.Show("Geen geldige invoer");
+            }
+            finally
+            {
+                btnPlaatsShowAllLocations_Click(sender, e);
+            }
         }
 
         private void btnPlaatsShowAllLocations_Click(object sender, EventArgs e)
         {
-
+            dataGridView.DataSource = null;
+            dataGridView.Refresh();
+            List<Location> showAllLocations = new List<Location>();
+            showAllLocations = superManager.ShowLocations();
+            dataGridView.DataSource = showAllLocations;
         }
 
         private void dataGridView_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -257,11 +323,8 @@ namespace EventManagementSystem
                         tbMedewerkerRoleID.Text = enumID.ToString();
                         cbMedewerkersRole.SelectedIndex = cbMedewerkersRole.FindStringExact(row.Cells["Role"].Value.ToString());
                     }
-                    
-
-                   
                 }
-                if (tabControl.SelectedTab == tabControl.TabPages["tpEvents"])
+                else if (tabControl.SelectedTab == tabControl.TabPages["tpEvents"])
                 {
                     tbEventEventID.Text = row.Cells["ID"].Value.ToString();
                     tbEventLocatie.Text = row.Cells["Location"].Value.ToString();
@@ -269,6 +332,13 @@ namespace EventManagementSystem
                     tbEventEindDatum.Text = row.Cells["EndDate"].Value.ToString();
                     tbEventBeschrijving.Text = row.Cells["Description"].Value.ToString();
                     tbEventPrijs.Text = row.Cells["AdmissionFee"].Value.ToString();
+                }
+                else
+                {
+                    tbPlaatsLocationID.Text = row.Cells["LOCATIONTYPEID"].Value.ToString();
+                    tbPlaatsName.Text = row.Cells["NAME"].Value.ToString();
+                    tbPlaatsDescription.Text = row.Cells["DESCRIPTION"].Value.ToString();
+                    tbPlaatsPrice.Text = row.Cells["PRICE"].Value.ToString(); 
                 }
